requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.0
pydantic==2.10.4
httpx==0.28.1
feedparser==6.0.11
apscheduler==3.10.4
python-multipart==0.0.20
aiofiles==24.1.0
Pillow==11.1.0
jinja2==3.1.5
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
filelock==3.16.1
python-dotenv==1.0.1
